requirements.txt
pyrogram
tgcrypto
pymongo
python-dotenv
flask
requests